chore: split pre-processing into a modular structure
- Refactored pre_processing.py into a package to improve maintainability and readability.
- Preserves the existing API for backward compatibility.
- No changes required for external imports or usage.
